rename package to mit_semseg
