+++ eval_multipro.py
@@ -10,12 +10,12 @@ import torch
 import torch.nn as nn
 from scipy.io import loadmat
 # Our libs
-from csail_semseg.config import cfg
-from csail_semseg.dataset import ValDataset
-from csail_semseg.models import ModelBuilder, SegmentationModule
-from csail_semseg.utils import AverageMeter, colorEncode, accuracy, intersectionAndUnion, parse_devices, setup_logger
-from csail_semseg.lib.nn import user_scattered_collate, async_copy_to
-from csail_semseg.lib.utils import as_numpy
+from mit_semseg.config import cfg
+from mit_semseg.dataset import ValDataset
+from mit_semseg.models import ModelBuilder, SegmentationModule
+from mit_semseg.utils import AverageMeter, colorEncode, accuracy, intersectionAndUnion, parse_devices, setup_logger
+from mit_semseg.lib.nn import user_scattered_collate, async_copy_to
+from mit_semseg.lib.utils import as_numpy
 from PIL import Image
 from tqdm import tqdm
