+++ mit_semseg/models/hrnet.py
@@ -8,7 +8,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from .utils import load_url
-from csail_semseg.lib.nn import SynchronizedBatchNorm2d
+from mit_semseg.lib.nn import SynchronizedBatchNorm2d
 
 BatchNorm2d = SynchronizedBatchNorm2d
 BN_MOMENTUM = 0.1
