+++ mit_semseg/models/mobilenet.py
@@ -6,7 +6,7 @@ https://github.com/tonylins/pytorch-mobilenet-v2
 import torch.nn as nn
 import math
 from .utils import load_url
-from csail_semseg.lib.nn import SynchronizedBatchNorm2d
+from mit_semseg.lib.nn import SynchronizedBatchNorm2d
 
 BatchNorm2d = SynchronizedBatchNorm2d
