+++ setup.py
@@ -4,9 +4,9 @@ with open('README.md', 'r') as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name='csail_semseg',
+    name='mit_semseg',
     version='1.0.0',
-    author='CSAIL',
+    author='MIT CSAIL',
     description='Pytorch implementation for Semantic Segmentation/Scene Parsing on MIT ADE20K dataset',
     long_description=long_description,
     long_description_content_type='text/markdown',
