+++ test.py
@@ -9,14 +9,14 @@ import torch.nn as nn
 from scipy.io import loadmat
 import csv
 # Our libs
-from csail_semseg.dataset import TestDataset
-from csail_semseg.models import ModelBuilder, SegmentationModule
-from csail_semseg.utils import colorEncode, find_recursive, setup_logger
-from csail_semseg.lib.nn import user_scattered_collate, async_copy_to
-from csail_semseg.lib.utils import as_numpy
+from mit_semseg.dataset import TestDataset
+from mit_semseg.models import ModelBuilder, SegmentationModule
+from mit_semseg.utils import colorEncode, find_recursive, setup_logger
+from mit_semseg.lib.nn import user_scattered_collate, async_copy_to
+from mit_semseg.lib.utils import as_numpy
 from PIL import Image
 from tqdm import tqdm
-from csail_semseg.config import cfg
+from mit_semseg.config import cfg
 
 colors = loadmat('data/color150.mat')['colors']
 names = {}
